requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
OTXv2==1.5.12
publicsuffixlist==0.10.0.20231214
pymisp==2.4.182
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
urllib3==2.1.0
wrapt==1.16.0